requirements.txt
numpy
pandas
yfinance
ensure
python-box
PyYAML
xgboost 
statsmodels 
scikit-learn 